Implements plugin error reporting using sentry.io (#37)
`SentrySdkInitializer` is used to initially set all the values that are error independent.

Sadly the Sentry SDK doesn't provide any transparent way to know when the report has been sent. So I had to use the class `ThreadSync` and `SentrySubmissionResult` to track when a result is set by the sdk and lock the background thread until then.

Information that is collected and sent to Sentry.io:
* Full StackTrace
* Plugin version
* IDE Name and Version (including exact build *number*)
* OS Name and Version
* JDK Vendor and Version (shouldn't be interesting to us as we have the IDE info but just in case)
* Last actionId (provided by `IdeaLogger#ourLastActionId`)

Attachments are currently not being sent as I am not sure about file size limits and other limitations we would be encountering in sentry.io.
But this would be achieved using `Sentry#withScope`, setting the attachments on the scope and have the `Sentry#captureEvent` call inside the scope-callback.