Make SentrySubmissionResult class package private
The result itself is never accessed or used by the error reported.
That's why I decided to throw an `UnsupportedOperationException` and not store the result somewhere.
It is given to the Consumer once and after that the result object is basically useless as the consumer gets set to null too.


https://github.com/php-perfect/ddev-intellij-plugin/pull/60#discussion_r883502010